requirements.txt
# ── HuggingFace Stack ──────────────────────────────────────
transformers==4.47.0
tokenizers==0.21.0
datasets==3.1.0
accelerate==1.2.1
huggingface-hub==0.26.2
safetensors==0.4.5

# ── Audio / Video ──────────────────────────────────────────
moviepy==1.0.3
imageio==2.34.2
imageio-ffmpeg==0.5.1
librosa==0.10.2
soundfile==0.12.1
audioread==3.0.1
opencv-python==4.10.0.84
ffmpeg-python==0.2.0

# ── Numerical / ML Core ────────────────────────────────────
# IMPORTANT: numpy MUST be <2.0 (facenet-pytorch hard requirement)
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
pandas==2.2.3
# IMPORTANT: Pillow must be ==10.2.0 (only version in >=10.2.0,<10.3.0)
Pillow==10.2.0

# ── Visualization ──────────────────────────────────────────
matplotlib==3.9.2
seaborn==0.13.2

# ── Utilities ──────────────────────────────────────────────
tqdm==4.67.1
regex==2024.9.11
filelock==3.20.0
packaging==24.1
psutil==6.1.0

# ── MLflow ─────────────────────────────────────────────────
mlflow==2.18.0
sqlalchemy==2.0.36
alembic==1.14.0
Flask==3.0.3
waitress==3.0.1; sys_platform == "win32"

# ── Decorator/misc for moviepy v1 ──────────────────────────
decorator==4.4.2
proglog==0.1.10
